# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lawyers Files",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://lawyers-files.otterscompany.com"
        }
    ],
    "paths": {
...
        "/api/v1/statements": {
            "get": {
                "tags": [
                    "Statements"
                ],
                "summary": "Get statements",
                "description": "Returns all statements",
                "operationId": "indexStatement",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all statements.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Statements"
                ],
                "summary": "Register statement",
                "description": "Create a new statement",
                "operationId": "storeStatement",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of statement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description of statement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of statement (1=Case, 2=Statement)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "1",
                                "2"
                            ]
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registered statement.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/statements/{id}": {
            "get": {
                "tags": [
                    "Statements"
                ],
                "summary": "Get statement",
                "description": "Returns a single statement",
                "operationId": "showStatement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show statement.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "No results found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Statements"
                ],
                "summary": "Update statement",
                "description": "Update a single statement",
                "operationId": "updateStatement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of statement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description of statement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of statement (1=Case, 2=Statement)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "1",
                                "2"
                            ]
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update statement.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Statements"
                ],
                "summary": "Delete statement",
                "description": "Delete a single statement",
                "operationId": "destroyStatement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete statement.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "No results found."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}